feat: use color class instead of hook
Created a color class that can be used as a value to the color property (in addition to ColorInput) and is the value provided to onChange. This makes it easier for users to get the color in whatever format they want instead of just providing hex.
Removed the useColorState hook as all of the color conversion logic can be provided without needing a react hook. This also allows users to decide how ever they think is best to maintain state for their application.